fix(db): separate utils for runtime (#10443)
* fix(db): separate utils for runtime

* add changeset

* use safeFetch from runtime utils
