+++ .changeset/few-teachers-impress.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/db": patch
+---
+
+Fixes an issue where `astro:db` could not be used in serverless environments.
