+++ packages/db/src/core/cli/commands/link/index.ts
@@ -7,7 +7,8 @@ import ora from 'ora';
 import prompts from 'prompts';
 import { MISSING_SESSION_ID_ERROR } from '../../../errors.js';
 import { PROJECT_ID_FILE, getSessionIdFromFile } from '../../../tokens.js';
-import { type Result, getAstroStudioUrl, safeFetch } from '../../../utils.js';
+import { type Result, getAstroStudioUrl } from '../../../utils.js';
+import { safeFetch } from '../../../../runtime/utils.js';
 
 export async function cmd() {
 	const sessionToken = await getSessionIdFromFile();
