+++ packages/db/src/core/cli/commands/push/index.ts
@@ -3,7 +3,8 @@ import type { Arguments } from 'yargs-parser';
 import { MIGRATION_VERSION } from '../../../consts.js';
 import { getManagedAppTokenOrExit } from '../../../tokens.js';
 import { type DBConfig, type DBSnapshot } from '../../../types.js';
-import { type Result, getRemoteDatabaseUrl, safeFetch } from '../../../utils.js';
+import { type Result, getRemoteDatabaseUrl } from '../../../utils.js';
+import { safeFetch } from '../../../../runtime/utils.js';
 import {
 	createCurrentSnapshot,
 	createEmptySnapshot,
